Move the graph layer and job submission into MereRunRelayKit
Extract the workflow graph documents, node registry, validator,
preflight diagnostic types, bundle materializer, and relay submission
out of the CLI target so any client can materialize and submit
byte-identical job bundles. Environment-specific lookups — default
model ids per node kind, managed-catalog provenance, plugin provider
discovery, and discovered plugin nodes — are injected through
WorkflowMaterializationEnvironment: the CLI supplies its catalog and
discovery (.cli), while clients without a local runtime use .portable,
which knows no plugin nodes and records identity-only provenance, so
they always pin models explicitly.

The CLI keeps its exact command surface through wrapper overloads and
converts RelayClientError at every graph subcommand boundary, so
messages, usage output, and exit codes are unchanged. swiftlint
--strict, swift build, and the full test suite (2912 tests, 0 failures)
pass.
