+++ Sources/MereRunCLI/Commands/TextChatCommand.swift
@@ -1,4 +1,5 @@
 import ArgumentParser
+import MereRunRelayKit
 import Foundation
 import MereRunCore
 #if canImport(Darwin)
