--- Sources/MereRunCLI/Support/MereRunCLIVersion.swift
@@ -1,3 +1,0 @@
-enum MereRunCLIVersion {
-    static let current = "0.40.1"
-}
