+++ Sources/MereRunCLI/Support/ModelPullPreflight.swift
@@ -1,4 +1,5 @@
 import Foundation
+import MereRunRelayKit
 import MereRunCore
 
 struct ModelPullPreflightInput {
