+++ Sources/MereRunCLI/Support/RelayKitCLIBridge.swift
@@ -1,7 +1,7 @@
 import ArgumentParser
+import MereRunRelayKit
 import Foundation
 import MereRunCore
-import MereRunRelayKit
 
 // MereRunRelayKit owns the relay client, executor profiles, and workflow wire
 // types; it is platform-neutral, so the CLI supplies its storage locations
