+++ Sources/MereRunCLI/Support/SFXVideoPreflight.swift
@@ -1,4 +1,5 @@
 import Foundation
+import MereRunRelayKit
 import MereRunCore
 
 struct SFXVideoPreflightInput {
