+++ Sources/MereRunCLI/Support/VideoGenerationPreflight.swift
@@ -1,4 +1,5 @@
 import Foundation
+import MereRunRelayKit
 import ArgumentParser
 import MereRunContract
 import MereRunCore
