+++ Sources/MereRunCLI/Support/WorkflowJobBundle.swift
@@ -1,151 +1,45 @@
 import ArgumentParser
+import MereRunRelayKit
 import Crypto
 import Foundation
 import MereRunCore
-import MereRunRelayKit
 
-struct WorkflowBundleMaterialization: Equatable {
-    let directory: URL
-    let graph: WorkflowGraphDocument
-    let inputs: WorkflowInputsDocument
-    let assets: WorkflowAssetManifest
-    let job: WorkflowJobManifest
-}
+// The materializer and bundle helpers live in MereRunRelayKit and take a
+// WorkflowMaterializationEnvironment. These wrappers keep the CLI's original
+// signatures, supplying the managed model catalog, per-command default model
+// ids, and discovered plugin providers only the CLI can know.
 
-func verifiedPortableWorkflowBundle(
-    at directory: URL,
-    fileManager: FileManager = .default
-) throws -> WorkflowBundleMaterialization {
-    let directory = directory.standardizedFileURL
-    for filename in [
-        WorkflowJobManifest.filename,
-        "graph.json",
-        "inputs.json",
-        WorkflowAssetManifest.filename,
-    ] {
-        let document = directory.appendingPathComponent(filename)
-        let values = try document.resourceValues(forKeys: [.isRegularFileKey, .isSymbolicLinkKey])
-        guard fileManager.fileExists(atPath: document.path),
-              values.isRegularFile == true,
-              values.isSymbolicLink != true else {
-            throw ValidationError("Workflow bundle document is missing or unsafe: \(filename)")
-        }
-    }
-    let graph = try WorkflowGraphDocument.load(from: directory.appendingPathComponent("graph.json"))
-    let inputs = try WorkflowInputsDocument.load(from: directory.appendingPathComponent("inputs.json"))
-    let assets = try WorkflowBundleCodec.decoder().decode(
-        WorkflowAssetManifest.self,
-        from: Data(contentsOf: directory.appendingPathComponent(WorkflowAssetManifest.filename))
-    )
-    let job = try WorkflowBundleCodec.decoder().decode(
-        WorkflowJobManifest.self,
-        from: Data(contentsOf: directory.appendingPathComponent(WorkflowJobManifest.filename))
-    )
-    guard job.contractVersion == WorkflowJobManifest.contractVersion else {
-        throw ValidationError("Unsupported workflow job contract '\(job.contractVersion)'.")
-    }
-    guard assets.schemaVersion == 1 else {
-        throw ValidationError("Unsupported workflow asset manifest version '\(assets.schemaVersion)'.")
-    }
-    guard try WorkflowBundleCodec.hash(graph) == job.graphFingerprint else {
-        throw ValidationError("Workflow graph fingerprint does not match job.json.")
-    }
-    let inputFingerprint = try WorkflowBundleCodec.hash(WorkflowPortableInputFingerprint(inputs: inputs, assets: assets))
-    guard inputFingerprint == job.inputFingerprint else {
-        throw ValidationError("Workflow input fingerprint does not match job.json.")
-    }
-    let validation = WorkflowGraphValidator.validate(graph: graph, inputs: inputs)
-    guard validation.status != .blocked else {
-        throw ValidationError(validation.diagnostics.map(\.message).joined(separator: " "))
-    }
-    let assetRoot = directory.appendingPathComponent("assets/sha256", isDirectory: true)
-    for entry in assets.groups.flatMap(\.entries) {
-        guard isPortableWorkflowPath(entry.path), isWorkflowSHA256(entry.digest) else {
-            throw ValidationError("Workflow asset manifest contains an invalid path or digest.")
-        }
-        let source = assetRoot.appendingPathComponent(entry.digest)
-        let values = try source.resourceValues(forKeys: [.isRegularFileKey, .isSymbolicLinkKey])
-        guard values.isRegularFile == true,
-              values.isSymbolicLink != true,
-              try ModelArtifactPin.fileByteCount(source) == entry.sizeBytes,
-              try ModelArtifactPin.fileSHA256(source) == entry.digest else {
-            throw ValidationError("Workflow asset digest verification failed for '\(entry.path)'.")
-        }
-    }
-    return WorkflowBundleMaterialization(
-        directory: directory,
-        graph: graph,
-        inputs: inputs,
-        assets: assets,
-        job: job
-    )
-}
-
-func copyPortableWorkflowBundle(
-    _ bundle: WorkflowBundleMaterialization,
-    to destination: URL,
-    fileManager: FileManager = .default
-) throws {
-    let destination = destination.standardizedFileURL
-    if fileManager.fileExists(atPath: destination.path),
-       !(try fileManager.contentsOfDirectory(atPath: destination.path)).isEmpty {
-        throw ValidationError("Workflow run directory must be empty before bundle submission.")
-    }
-    try fileManager.createDirectory(at: destination, withIntermediateDirectories: true)
-    for filename in [
-        WorkflowJobManifest.filename,
-        "graph.json",
-        "inputs.json",
-        WorkflowAssetManifest.filename,
-    ] {
-        try fileManager.copyItem(
-            at: bundle.directory.appendingPathComponent(filename),
-            to: destination.appendingPathComponent(filename)
-        )
-    }
-    let destinationAssets = destination.appendingPathComponent("assets/sha256", isDirectory: true)
-    try fileManager.createDirectory(at: destinationAssets, withIntermediateDirectories: true)
-    for digest in Set(bundle.assets.groups.flatMap(\.entries).map(\.digest)).sorted() {
-        try fileManager.copyItem(
-            at: bundle.directory.appendingPathComponent("assets/sha256/\(digest)"),
-            to: destinationAssets.appendingPathComponent(digest)
+extension WorkflowMaterializationEnvironment {
+    static var cli: WorkflowMaterializationEnvironment {
+        WorkflowMaterializationEnvironment(
+            mereRunVersion: MereRunCLIVersion.current,
+            pluginNodes: WorkflowGraphProviderRegistry.discoveredCatalog().nodes,
+            defaultModelID: { nodeKind in
+                switch nodeKind {
+                case "image.train-lora": ImageTrainLoRA.defaultManagedModelID.rawValue
+                case "image.generate": ImageGenerate.defaultManagedModelID.rawValue
+                case "vision.ground": VisionGround.defaultManagedModelID.rawValue
+                case "vision.segment", "vision.track": VisionSegment.defaultManagedModelID.rawValue
+                case "video.generate": ModelResolver.ModelID.ltxVideo23AVMLX.rawValue
+                default: nil
+                }
+            },
+            modelProvenance: { modelID in
+                let spec = ManagedModelCatalog.spec(for: modelID)
+                return .fromCatalogIdentity(
+                    id: modelID,
+                    repository: spec?.upstreamRepoId,
+                    revision: spec?.upstreamRevision
+                )
+            },
+            providerRequirement: { providerID in
+                WorkflowGraphProviderRegistry.discoveredCatalog().provider(id: providerID)?.requirement
+            }
         )
     }
 }
 
-func requireSeparateWorkflowDirectories(bundle: URL, run: URL) throws {
-    let bundle = bundle.standardizedFileURL.path
-    let run = run.standardizedFileURL.path
-    guard bundle != run,
-          !run.hasPrefix(bundle + "/"),
-          !bundle.hasPrefix(run + "/") else {
-        throw ValidationError("Immutable bundle and run directories must be separate and non-nested.")
-    }
-}
-
-private func isPortableWorkflowPath(_ path: String) -> Bool {
-    !path.isEmpty
-        && !path.hasPrefix("/")
-        && path.split(separator: "/", omittingEmptySubsequences: false).allSatisfy {
-            $0 != "." && $0 != ".." && !$0.isEmpty
-        }
-}
-
-private func isWorkflowSHA256(_ value: String) -> Bool {
-    value.count == 64 && value.utf8.allSatisfy { byte in
-        (byte >= 48 && byte <= 57) || (byte >= 97 && byte <= 102)
-    }
-}
-
-struct WorkflowBundleMaterializer {
-    let graph: WorkflowGraphDocument
-    let suppliedInputs: WorkflowInputsDocument
-    let destination: URL
-    let fileManager: FileManager
-    let now: () -> Date
-    let jobID: () -> UUID
-    let seed: () -> Int64
-
+extension WorkflowBundleMaterializer {
     init(
         graph: WorkflowGraphDocument,
         suppliedInputs: WorkflowInputsDocument,
@@ -155,432 +49,30 @@ struct WorkflowBundleMaterializer {
         jobID: @escaping () -> UUID = UUID.init,
         seed: @escaping () -> Int64 = { Int64.random(in: 0...Int64.max) }
     ) {
-        self.graph = graph
-        self.suppliedInputs = suppliedInputs
-        self.destination = destination.standardizedFileURL
-        self.fileManager = fileManager
-        self.now = now
-        self.jobID = jobID
-        self.seed = seed
-    }
-
-    func materialize() throws -> WorkflowBundleMaterialization {
-        let validation = WorkflowGraphValidator.validate(graph: graph, inputs: suppliedInputs)
-        guard validation.status != .blocked else {
-            let messages = validation.diagnostics.filter { $0.severity == .blocker }.map(\.message)
-            throw ValidationError(messages.joined(separator: " "))
-        }
-        try prepareDestination()
-
-        let sourceGraphFingerprint = try WorkflowBundleCodec.hash(graph)
-        let sourceInputFingerprint = try WorkflowBundleCodec.hash(suppliedInputs)
-        let resolvedInputs = resolveDefaults()
-        let resolvedGraph = resolveSeeds()
-        let assetsDirectory = destination
-            .appendingPathComponent("assets", isDirectory: true)
-            .appendingPathComponent("sha256", isDirectory: true)
-        try fileManager.createDirectory(at: assetsDirectory, withIntermediateDirectories: true)
-
-        var portableValues = resolvedInputs.values
-        var groups: [WorkflowAssetGroup] = []
-        for name in resolvedGraph.inputs.keys.sorted() {
-            guard let definition = resolvedGraph.inputs[name],
-                  definition.type == .asset || definition.type == .assetDirectory,
-                  let sourcePath = resolvedInputs.values[name]?.stringValue else {
-                continue
-            }
-            let sourceURL = URL(fileURLWithPath: sourcePath).standardizedFileURL
-            let group = try materializeAsset(name: name, definition: definition, sourceURL: sourceURL, assetsDirectory: assetsDirectory)
-            groups.append(group)
-            portableValues[name] = .string("asset://\(name)")
-        }
-
-        let portableGraph = try materializeNodeAssetConstants(
-            graph: resolvedGraph,
-            assetsDirectory: assetsDirectory,
-            groups: &groups
+        self.init(
+            graph: graph,
+            suppliedInputs: suppliedInputs,
+            destination: destination,
+            environment: .cli,
+            fileManager: fileManager,
+            now: now,
+            jobID: jobID,
+            seed: seed
         )
-        let portableInputs = WorkflowInputsDocument(values: portableValues)
-        let assets = WorkflowAssetManifest(schemaVersion: 1, groups: groups)
-        let graphFingerprint = try WorkflowBundleCodec.hash(portableGraph)
-        let inputFingerprint = try WorkflowBundleCodec.hash(
-            WorkflowPortableInputFingerprint(inputs: portableInputs, assets: assets)
-        )
-        let job = WorkflowJobManifest(
-            contractVersion: WorkflowJobManifest.contractVersion,
-            jobID: jobID().uuidString.lowercased(),
-            createdAt: now(),
-            graphFingerprint: graphFingerprint,
-            inputFingerprint: inputFingerprint,
-            sourceGraphFingerprint: sourceGraphFingerprint,
-            sourceInputFingerprint: sourceInputFingerprint,
-            requirements: requirements(graph: portableGraph, resolvedInputs: resolvedInputs),
-            outputs: portableGraph.outputs.keys.sorted().compactMap { name in
-                guard case .reference(let reference)? = portableGraph.outputs[name] else { return nil }
-                return WorkflowJobOutput(name: name, reference: reference)
-            }
-        )
-
-        try WorkflowBundleCodec.write(portableGraph, to: destination.appendingPathComponent("graph.json"))
-        try WorkflowBundleCodec.write(portableInputs, to: destination.appendingPathComponent("inputs.json"))
-        try WorkflowBundleCodec.write(assets, to: destination.appendingPathComponent(WorkflowAssetManifest.filename))
-        try WorkflowBundleCodec.write(job, to: destination.appendingPathComponent(WorkflowJobManifest.filename))
-
-        return WorkflowBundleMaterialization(
-            directory: destination,
-            graph: portableGraph,
-            inputs: portableInputs,
-            assets: assets,
-            job: job
-        )
-    }
-
-    private func prepareDestination() throws {
-        var isDirectory: ObjCBool = false
-        if fileManager.fileExists(atPath: destination.path, isDirectory: &isDirectory) {
-            guard isDirectory.boolValue else {
-                throw ValidationError("Job bundle destination is not a directory: \(destination.path)")
-            }
-            let existing = try fileManager.contentsOfDirectory(atPath: destination.path)
-            guard existing.isEmpty else {
-                throw ValidationError("Job bundle destination is not empty: \(destination.path)")
-            }
-        } else {
-            try fileManager.createDirectory(at: destination, withIntermediateDirectories: true)
-        }
-    }
-
-    private func resolveDefaults() -> WorkflowInputsDocument {
-        var values = suppliedInputs.values
-        for (name, definition) in graph.inputs where values[name] == nil {
-            values[name] = definition.defaultValue
-        }
-        return WorkflowInputsDocument(values: values)
-    }
-
-    private func resolveSeeds() -> WorkflowGraphDocument {
-        let nodes = graph.nodes.map { node -> WorkflowNode in
-            guard let catalog = WorkflowNodeRegistry.entry(for: node) else { return node }
-            var arguments = node.arguments
-            for field in catalog.inputs where arguments[field.name] == nil {
-                arguments[field.name] = field.defaultValue
-            }
-            if catalog.inputs.contains(where: { $0.name == "seed" }), arguments["seed"] == nil {
-                arguments["seed"] = .integer(seed())
-            }
-            return WorkflowNode(
-                id: node.id,
-                kind: node.kind,
-                provider: node.provider,
-                arguments: arguments,
-                dependsOn: node.dependsOn,
-                execution: node.execution
-            )
-        }
-        return WorkflowGraphDocument(
-            schemaVersion: graph.schemaVersion,
-            kind: graph.kind,
-            name: graph.name,
-            inputs: graph.inputs,
-            nodes: nodes,
-            outputs: graph.outputs,
-            execution: graph.execution,
-            metadata: graph.metadata
-        )
-    }
-
-    private func materializeAsset(
-        name: String,
-        definition: WorkflowInputDefinition,
-        sourceURL: URL,
-        assetsDirectory: URL
-    ) throws -> WorkflowAssetGroup {
-        let resourceValues = try sourceURL.resourceValues(forKeys: [.isRegularFileKey, .isDirectoryKey, .isSymbolicLinkKey])
-        if resourceValues.isSymbolicLink == true {
-            throw ValidationError("Workflow asset '\(name)' cannot be a symbolic link: \(sourceURL.path)")
-        }
-        switch definition.type {
-        case .asset:
-            guard resourceValues.isRegularFile == true else {
-                throw ValidationError("Workflow asset '\(name)' is not a regular file: \(sourceURL.path)")
-            }
-            return WorkflowAssetGroup(
-                name: name,
-                kind: .asset,
-                entries: [try materializeFile(sourceURL, relativePath: sourceURL.lastPathComponent, assetsDirectory: assetsDirectory)]
-            )
-        case .assetDirectory:
-            guard resourceValues.isDirectory == true else {
-                throw ValidationError("Workflow asset directory '\(name)' was not found: \(sourceURL.path)")
-            }
-            let confinedRoot = sourceURL.resolvingSymlinksInPath()
-            let entries = try directoryFiles(confinedRoot).map { file in
-                try materializeFile(file.url, relativePath: file.relativePath, assetsDirectory: assetsDirectory)
-            }
-            guard !entries.isEmpty else {
-                throw ValidationError("Workflow asset directory '\(name)' is empty: \(sourceURL.path)")
-            }
-            return WorkflowAssetGroup(name: name, kind: .assetDirectory, entries: entries)
-        default:
-            throw ValidationError("Input '\(name)' is not an asset input.")
-        }
-    }
-
-    private func materializeNodeAssetConstants(
-        graph: WorkflowGraphDocument,
-        assetsDirectory: URL,
-        groups: inout [WorkflowAssetGroup]
-    ) throws -> WorkflowGraphDocument {
-        let nodes = try graph.nodes.map { node -> WorkflowNode in
-            guard let catalog = WorkflowNodeRegistry.entry(for: node) else { return node }
-            var arguments = node.arguments
-            for field in catalog.inputs {
-                guard let value = arguments[field.name] else { continue }
-                switch field.type {
-                case .asset:
-                    arguments[field.name] = try portableNodeAsset(
-                        value,
-                        kind: .asset,
-                        key: "\(node.id).\(field.name).0",
-                        assetsDirectory: assetsDirectory,
-                        groups: &groups
-                    )
-                case .assetDirectory:
-                    arguments[field.name] = try portableNodeAsset(
-                        value,
-                        kind: .assetDirectory,
-                        key: "\(node.id).\(field.name).0",
-                        assetsDirectory: assetsDirectory,
-                        groups: &groups
-                    )
-                case .assetArray, .assetCollection:
-                    guard case .array(let values) = value else { continue }
-                    arguments[field.name] = .array(try values.enumerated().map { index, nested in
-                        try portableNodeAsset(
-                            nested,
-                            kind: .asset,
-                            key: "\(node.id).\(field.name).\(index)",
-                            assetsDirectory: assetsDirectory,
-                            groups: &groups
-                        )
-                    })
-                default:
-                    break
-                }
-            }
-            return WorkflowNode(
-                id: node.id,
-                kind: node.kind,
-                provider: node.provider,
-                arguments: arguments,
-                dependsOn: node.dependsOn,
-                execution: node.execution
-            )
-        }
-        return WorkflowGraphDocument(
-            schemaVersion: graph.schemaVersion,
-            kind: graph.kind,
-            name: graph.name,
-            inputs: graph.inputs,
-            nodes: nodes,
-            outputs: graph.outputs,
-            execution: graph.execution,
-            metadata: graph.metadata
-        )
-    }
-
-    private func portableNodeAsset(
-        _ value: WorkflowValue,
-        kind: WorkflowInputType,
-        key: String,
-        assetsDirectory: URL,
-        groups: inout [WorkflowAssetGroup]
-    ) throws -> WorkflowValue {
-        guard case .string(let path) = value else { return value }
-        guard !path.hasPrefix("asset://") else {
-            throw ValidationError("asset:// is reserved for materialized workflow assets.")
-        }
-        let digest = SHA256.hash(data: Data(key.utf8)).prefix(16).map { String(format: "%02x", $0) }.joined()
-        let name = "asset-\(digest)"
-        guard !groups.contains(where: { $0.name == name }) else {
-            throw ValidationError("Workflow asset identifier collision for '\(key)'.")
-        }
-        let definition = WorkflowInputDefinition(
-            type: kind,
-            required: true,
-            defaultValue: nil,
-            values: nil,
-            contentTypes: nil
-        )
-        let group = try materializeAsset(
-            name: name,
-            definition: definition,
-            sourceURL: URL(fileURLWithPath: path).standardizedFileURL,
-            assetsDirectory: assetsDirectory
-        )
-        groups.append(group)
-        return .string("asset://\(name)")
-    }
-
-    private func directoryFiles(_ root: URL) throws -> [(url: URL, relativePath: String)] {
-        var files: [(url: URL, relativePath: String)] = []
-        for relativePath in try fileManager.subpathsOfDirectory(atPath: root.path).sorted() {
-            if relativePath.split(separator: "/").contains(where: { $0.hasPrefix(".") }) { continue }
-            let url = root.appendingPathComponent(relativePath)
-            let values = try url.resourceValues(forKeys: [.isRegularFileKey, .isSymbolicLinkKey])
-            if values.isSymbolicLink == true {
-                throw ValidationError("Workflow asset directories cannot contain symbolic links: \(url.path)")
-            }
-            if values.isRegularFile == true {
-                files.append((url, relativePath))
-            }
-        }
-        return files
-    }
-
-    private func materializeFile(
-        _ sourceURL: URL,
-        relativePath: String,
-        assetsDirectory: URL
-    ) throws -> WorkflowAssetEntry {
-        guard !relativePath.hasPrefix("/"), !relativePath.split(separator: "/").contains("..") else {
-            throw ValidationError("Workflow asset path escapes its declared root: \(relativePath)")
-        }
-        let digest = try ModelArtifactPin.fileSHA256(sourceURL)
-        let size = try ModelArtifactPin.fileByteCount(sourceURL)
-        let destinationURL = assetsDirectory.appendingPathComponent(digest)
-        if !fileManager.fileExists(atPath: destinationURL.path) {
-            try fileManager.copyItem(at: sourceURL, to: destinationURL)
-        }
-        return WorkflowAssetEntry(
-            path: relativePath,
-            digest: digest,
-            sizeBytes: size,
-            contentType: contentType(for: sourceURL)
-        )
-    }
-
-    private func requirements(
-        graph: WorkflowGraphDocument,
-        resolvedInputs: WorkflowInputsDocument
-    ) -> WorkflowJobRequirements {
-        var modelIDs = Set(graph.nodes.compactMap { node -> String? in
-            if let value = node.arguments["model"] {
-                if let resolved = resolveStatic(value, inputs: resolvedInputs.values)?.stringValue {
-                    return resolved
-                }
-                return nil
-            }
-            switch node.kind {
-            case "image.train-lora": return ImageTrainLoRA.defaultManagedModelID.rawValue
-            case "image.generate": return ImageGenerate.defaultManagedModelID.rawValue
-            case "vision.ground": return VisionGround.defaultManagedModelID.rawValue
-            case "vision.segment", "vision.track": return VisionSegment.defaultManagedModelID.rawValue
-            case "video.generate": return ModelResolver.ModelID.ltxVideo23AVMLX.rawValue
-            default: return nil
-            }
-        })
-        for node in graph.nodes {
-            modelIDs.formUnion(WorkflowNodeRegistry.entry(for: node)?.requirements.modelIDs ?? [])
-        }
-        let modelProvenance = modelIDs.sorted().map { modelID -> WorkflowModelProvenance in
-            let spec = ManagedModelCatalog.spec(for: modelID)
-            let catalogIdentity = WorkflowManagedModelIdentity(
-                id: modelID,
-                repository: spec?.upstreamRepoId,
-                revision: spec?.upstreamRevision
-            )
-            return WorkflowModelProvenance(
-                id: modelID,
-                repository: spec?.upstreamRepoId,
-                revision: spec?.upstreamRevision,
-                catalogSHA256: (try? WorkflowBundleCodec.hash(catalogIdentity)) ?? "",
-                installManifestSHA256: nil
-            )
-        }
-        let providers = Array(Set(graph.nodes.compactMap { node -> WorkflowGraphProviderRequirement? in
-            guard node.resolvedProviderID != WorkflowNodeProviderIdentity.builtInID else { return nil }
-            return WorkflowGraphProviderRegistry.discoveredCatalog().provider(id: node.resolvedProviderID)?.requirement
-        })).sorted { $0.id < $1.id }
-        var acceptedBackends = Set(["cpu", "metal", "cuda", "rocm"])
-        var minimumMemory: Int64?
-        var minimumSystemMemory: Int64?
-        var minimumDisk: Int64?
-        var minimumCPUCores: Int?
-        var networkAccess = false
-        for node in graph.nodes {
-            guard let requirements = WorkflowNodeRegistry.entry(for: node)?.requirements else { continue }
-            if !requirements.acceleratorBackends.isEmpty {
-                acceptedBackends.formIntersection(requirements.acceleratorBackends)
-            }
-            if let nodeMinimum = requirements.minimumAcceleratorMemoryBytes {
-                minimumMemory = max(minimumMemory ?? 0, nodeMinimum)
-            }
-            if let nodeMinimum = requirements.minimumSystemMemoryBytes {
-                minimumSystemMemory = max(minimumSystemMemory ?? 0, nodeMinimum)
-            }
-            if let nodeMinimum = requirements.minimumDiskBytes {
-                minimumDisk = max(minimumDisk ?? 0, nodeMinimum)
-            }
-            if let nodeMinimum = requirements.minimumCPUCores {
-                minimumCPUCores = max(minimumCPUCores ?? 0, nodeMinimum)
-            }
-            networkAccess = networkAccess || requirements.networkAccess == true
-        }
-        let secretNames = Array(Set(
-            graph.nodes.flatMap { $0.arguments.values.flatMap(\.secretNames) }
-        )).sorted()
-        return WorkflowJobRequirements(
-            minimumMereRunVersion: MereRunCLIVersion.current,
-            nodeKinds: Array(Set(graph.nodes.map(\.kind))).sorted(),
-            modelIDs: modelIDs.sorted(),
-            models: modelProvenance,
-            providers: providers,
-            secretNames: secretNames,
-            acceleratorBackends: acceptedBackends.sorted(),
-            minimumAcceleratorMemoryBytes: minimumMemory,
-            minimumSystemMemoryBytes: minimumSystemMemory,
-            minimumDiskBytes: minimumDisk,
-            minimumCPUCores: minimumCPUCores,
-            networkAccess: networkAccess
-        )
-    }
-
-    private func resolveStatic(_ value: WorkflowValue, inputs: [String: WorkflowValue]) -> WorkflowValue? {
-        guard case .reference(let rawReference) = value,
-              let reference = try? WorkflowReference(rawReference),
-              case .input(let input) = reference.source else {
-            return value
-        }
-        return inputs[input]
-    }
-
-    private func contentType(for url: URL) -> String {
-        switch url.pathExtension.lowercased() {
-        case "png": "image/png"
-        case "jpg", "jpeg": "image/jpeg"
-        case "webp": "image/webp"
-        case "mp4": "video/mp4"
-        case "wav": "audio/wav"
-        case "tif", "tiff": "image/tiff"
-        case "json": "application/json"
-        case "txt": "text/plain"
-        case "safetensors": "application/x-safetensors"
-        default: "application/octet-stream"
-        }
     }
 }
 
-private struct WorkflowManagedModelIdentity: Codable {
-    let id: String
-    let repository: String?
-    let revision: String?
+func verifiedPortableWorkflowBundle(
+    at directory: URL,
+    fileManager: FileManager = .default
+) throws -> WorkflowBundleMaterialization {
+    try verifiedPortableWorkflowBundle(
+        at: directory,
+        fileManager: fileManager,
+        pluginNodes: WorkflowGraphProviderRegistry.discoveredCatalog().nodes
+    )
 }
 
-struct WorkflowPortableInputFingerprint: Codable {
-    let inputs: WorkflowInputsDocument
-    let assets: WorkflowAssetManifest
-}
 
 struct WorkflowInvocationOutput: Codable, Equatable, Sendable {
     let type: WorkflowPortType
