+++ Sources/MereRunRelayKit/WorkflowGraphDocuments.swift
@@ -1,16 +1,28 @@
-import ArgumentParser
 import Foundation
-import MereRunRelayKit
 
-typealias WorkflowInputType = WorkflowPortType
-typealias WorkflowFieldType = WorkflowPortType
+public typealias WorkflowInputType = WorkflowPortType
+public typealias WorkflowFieldType = WorkflowPortType
 
-struct WorkflowInputDefinition: Codable, Equatable, Sendable {
-    let type: WorkflowInputType
-    let required: Bool?
-    let defaultValue: WorkflowValue?
-    let values: [String]?
-    let contentTypes: [String]?
+public struct WorkflowInputDefinition: Codable, Equatable, Sendable {
+    public let type: WorkflowInputType
+    public let required: Bool?
+    public let defaultValue: WorkflowValue?
+    public let values: [String]?
+    public let contentTypes: [String]?
+
+    public init(
+        type: WorkflowInputType,
+        required: Bool?,
+        defaultValue: WorkflowValue?,
+        values: [String]?,
+        contentTypes: [String]?
+    ) {
+        self.type = type
+        self.required = required
+        self.defaultValue = defaultValue
+        self.values = values
+        self.contentTypes = contentTypes
+    }
 
     enum CodingKeys: String, CodingKey {
         case type
@@ -21,19 +33,19 @@ struct WorkflowInputDefinition: Codable, Equatable, Sendable {
     }
 }
 
-enum WorkflowNodeCachePolicy: String, Codable, Equatable, Sendable {
+public enum WorkflowNodeCachePolicy: String, Codable, Equatable, Sendable {
     case automatic = "auto"
     case never
     case refresh
 }
 
-struct WorkflowNodeExecutionPolicy: Codable, Equatable, Sendable {
-    static let maximumAttempts = 10
-    static let maximumTimeoutSeconds = 7 * 24 * 60 * 60
+public struct WorkflowNodeExecutionPolicy: Codable, Equatable, Sendable {
+    public static let maximumAttempts = 10
+    public static let maximumTimeoutSeconds = 7 * 24 * 60 * 60
 
-    let maxAttempts: Int?
-    let timeoutSeconds: Int?
-    let cache: WorkflowNodeCachePolicy?
+    public let maxAttempts: Int?
+    public let timeoutSeconds: Int?
+    public let cache: WorkflowNodeCachePolicy?
 
     enum CodingKeys: String, CodingKey {
         case maxAttempts = "max_attempts"
@@ -41,29 +53,29 @@ struct WorkflowNodeExecutionPolicy: Codable, Equatable, Sendable {
         case cache
     }
 
-    var resolvedMaxAttempts: Int { maxAttempts ?? 1 }
-    var resolvedCache: WorkflowNodeCachePolicy { cache ?? .automatic }
+    public var resolvedMaxAttempts: Int { maxAttempts ?? 1 }
+    public var resolvedCache: WorkflowNodeCachePolicy { cache ?? .automatic }
 }
 
-struct WorkflowGraphExecutionPolicy: Codable, Equatable, Sendable {
-    static let maximumParallelNodes = 32
+public struct WorkflowGraphExecutionPolicy: Codable, Equatable, Sendable {
+    public static let maximumParallelNodes = 32
 
-    let maxParallelNodes: Int?
+    public let maxParallelNodes: Int?
 
     enum CodingKeys: String, CodingKey {
         case maxParallelNodes = "max_parallel_nodes"
     }
 
-    var resolvedMaxParallelNodes: Int { maxParallelNodes ?? 1 }
+    public var resolvedMaxParallelNodes: Int { maxParallelNodes ?? 1 }
 }
 
-struct WorkflowNode: Codable, Equatable, Sendable {
-    let id: String
-    let kind: String
-    let provider: String?
-    let arguments: [String: WorkflowValue]
-    let dependsOn: [String]?
-    let execution: WorkflowNodeExecutionPolicy?
+public struct WorkflowNode: Codable, Equatable, Sendable {
+    public let id: String
+    public let kind: String
+    public let provider: String?
+    public let arguments: [String: WorkflowValue]
+    public let dependsOn: [String]?
+    public let execution: WorkflowNodeExecutionPolicy?
 
     enum CodingKeys: String, CodingKey {
         case id
@@ -74,7 +86,7 @@ struct WorkflowNode: Codable, Equatable, Sendable {
         case execution
     }
 
-    init(
+    public init(
         id: String,
         kind: String,
         provider: String? = nil,
@@ -90,23 +102,23 @@ struct WorkflowNode: Codable, Equatable, Sendable {
         self.execution = execution
     }
 
-    var resolvedProviderID: String {
+    public var resolvedProviderID: String {
         provider ?? WorkflowNodeProviderIdentity.builtInID
     }
 }
 
-struct WorkflowGraphDocument: Codable, Equatable, Sendable {
-    static let schemaVersion = 1
-    static let kind = "mere.run/workflow-graph"
+public struct WorkflowGraphDocument: Codable, Equatable, Sendable {
+    public static let schemaVersion = 1
+    public static let kind = "mere.run/workflow-graph"
 
-    let schemaVersion: Int
-    let kind: String
-    let name: String
-    let inputs: [String: WorkflowInputDefinition]
-    let nodes: [WorkflowNode]
-    let outputs: [String: WorkflowValue]
-    let execution: WorkflowGraphExecutionPolicy?
-    let metadata: [String: WorkflowValue]?
+    public let schemaVersion: Int
+    public let kind: String
+    public let name: String
+    public let inputs: [String: WorkflowInputDefinition]
+    public let nodes: [WorkflowNode]
+    public let outputs: [String: WorkflowValue]
+    public let execution: WorkflowGraphExecutionPolicy?
+    public let metadata: [String: WorkflowValue]?
 
     enum CodingKeys: String, CodingKey {
         case schemaVersion = "schema_version"
@@ -119,7 +131,7 @@ struct WorkflowGraphDocument: Codable, Equatable, Sendable {
         case metadata
     }
 
-    init(
+    public init(
         schemaVersion: Int,
         kind: String,
         name: String,
@@ -139,35 +151,35 @@ struct WorkflowGraphDocument: Codable, Equatable, Sendable {
         self.metadata = metadata
     }
 
-    static func load(from url: URL) throws -> WorkflowGraphDocument {
+    public static func load(from url: URL) throws -> WorkflowGraphDocument {
         try JSONDecoder().decode(WorkflowGraphDocument.self, from: Data(contentsOf: url))
     }
 }
 
-struct WorkflowInputsDocument: Codable, Equatable, Sendable {
-    let values: [String: WorkflowValue]
+public struct WorkflowInputsDocument: Codable, Equatable, Sendable {
+    public let values: [String: WorkflowValue]
 
-    init(values: [String: WorkflowValue]) {
+    public init(values: [String: WorkflowValue]) {
         self.values = values
     }
 
-    init(from decoder: Decoder) throws {
+    public init(from decoder: Decoder) throws {
         let container = try decoder.singleValueContainer()
         values = try container.decode([String: WorkflowValue].self)
     }
 
-    func encode(to encoder: Encoder) throws {
+    public func encode(to encoder: Encoder) throws {
         var container = encoder.singleValueContainer()
         try container.encode(values)
     }
 
-    static func load(from url: URL?) throws -> WorkflowInputsDocument {
+    public static func load(from url: URL?) throws -> WorkflowInputsDocument {
         guard let url else { return WorkflowInputsDocument(values: [:]) }
         return try JSONDecoder().decode(WorkflowInputsDocument.self, from: Data(contentsOf: url))
     }
 }
 
-enum WorkflowValueSchemaType: String, Codable, Equatable, Sendable {
+public enum WorkflowValueSchemaType: String, Codable, Equatable, Sendable {
     case string
     case integer
     case number
@@ -178,20 +190,20 @@ enum WorkflowValueSchemaType: String, Codable, Equatable, Sendable {
     case enumeration = "enum"
 }
 
-final class WorkflowValueSchema: Codable, @unchecked Sendable, Equatable {
-    let type: WorkflowValueSchemaType
-    let title: String?
-    let description: String?
-    let defaultValue: WorkflowValue?
-    let values: [String]?
-    let properties: [String: WorkflowValueSchema]?
-    let required: [String]?
-    let items: WorkflowValueSchema?
-    let additionalProperties: WorkflowValueSchema?
-    let minimum: Double?
-    let maximum: Double?
-    let step: Double?
-    let multiline: Bool?
+public final class WorkflowValueSchema: Codable, @unchecked Sendable, Equatable {
+    public let type: WorkflowValueSchemaType
+    public let title: String?
+    public let description: String?
+    public let defaultValue: WorkflowValue?
+    public let values: [String]?
+    public let properties: [String: WorkflowValueSchema]?
+    public let required: [String]?
+    public let items: WorkflowValueSchema?
+    public let additionalProperties: WorkflowValueSchema?
+    public let minimum: Double?
+    public let maximum: Double?
+    public let step: Double?
+    public let multiline: Bool?
 
     enum CodingKeys: String, CodingKey {
         case type
@@ -209,7 +221,7 @@ final class WorkflowValueSchema: Codable, @unchecked Sendable, Equatable {
         case multiline
     }
 
-    init(
+    public init(
         type: WorkflowValueSchemaType,
         title: String? = nil,
         description: String? = nil,
@@ -239,7 +251,7 @@ final class WorkflowValueSchema: Codable, @unchecked Sendable, Equatable {
         self.multiline = multiline
     }
 
-    static func == (lhs: WorkflowValueSchema, rhs: WorkflowValueSchema) -> Bool {
+    public static func == (lhs: WorkflowValueSchema, rhs: WorkflowValueSchema) -> Bool {
         lhs.type == rhs.type
             && lhs.title == rhs.title
             && lhs.description == rhs.description
@@ -256,21 +268,21 @@ final class WorkflowValueSchema: Codable, @unchecked Sendable, Equatable {
     }
 }
 
-struct WorkflowNodeField: Codable, Equatable, Sendable {
-    let name: String
-    let type: WorkflowFieldType
-    let required: Bool
-    let description: String?
-    let defaultValue: WorkflowValue?
-    let values: [String]?
-    let acceptedContentTypes: [String]?
-    let minimum: Double?
-    let maximum: Double?
-    let step: Double?
-    let multiline: Bool?
-    let secret: Bool?
-    let advanced: Bool?
-    let valueSchema: WorkflowValueSchema?
+public struct WorkflowNodeField: Codable, Equatable, Sendable {
+    public let name: String
+    public let type: WorkflowFieldType
+    public let required: Bool
+    public let description: String?
+    public let defaultValue: WorkflowValue?
+    public let values: [String]?
+    public let acceptedContentTypes: [String]?
+    public let minimum: Double?
+    public let maximum: Double?
+    public let step: Double?
+    public let multiline: Bool?
+    public let secret: Bool?
+    public let advanced: Bool?
+    public let valueSchema: WorkflowValueSchema?
 
     enum CodingKeys: String, CodingKey {
         case name
@@ -289,7 +301,7 @@ struct WorkflowNodeField: Codable, Equatable, Sendable {
         case valueSchema = "value_schema"
     }
 
-    init(
+    public init(
         name: String,
         type: WorkflowFieldType,
         required: Bool,
@@ -322,12 +334,12 @@ struct WorkflowNodeField: Codable, Equatable, Sendable {
     }
 }
 
-struct WorkflowNodeOutput: Codable, Equatable, Sendable {
-    let name: String
-    let type: WorkflowFieldType
-    let optional: Bool
-    let description: String?
-    let contentTypes: [String]
+public struct WorkflowNodeOutput: Codable, Equatable, Sendable {
+    public let name: String
+    public let type: WorkflowFieldType
+    public let optional: Bool
+    public let description: String?
+    public let contentTypes: [String]
 
     enum CodingKeys: String, CodingKey {
         case name
@@ -337,7 +349,7 @@ struct WorkflowNodeOutput: Codable, Equatable, Sendable {
         case contentTypes = "content_types"
     }
 
-    init(
+    public init(
         name: String,
         type: WorkflowFieldType,
         optional: Bool = false,
@@ -351,7 +363,7 @@ struct WorkflowNodeOutput: Codable, Equatable, Sendable {
         self.contentTypes = contentTypes
     }
 
-    init(from decoder: Decoder) throws {
+    public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         name = try container.decode(String.self, forKey: .name)
         type = try container.decode(WorkflowFieldType.self, forKey: .type)
@@ -360,7 +372,7 @@ struct WorkflowNodeOutput: Codable, Equatable, Sendable {
         contentTypes = try container.decodeIfPresent([String].self, forKey: .contentTypes) ?? []
     }
 
-    func encode(to encoder: Encoder) throws {
+    public func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(name, forKey: .name)
         try container.encode(type, forKey: .type)
@@ -372,14 +384,14 @@ struct WorkflowNodeOutput: Codable, Equatable, Sendable {
     }
 }
 
-struct WorkflowNodeRequirements: Codable, Equatable, Sendable {
-    let modelIDs: [String]
-    let acceleratorBackends: [String]
-    let minimumAcceleratorMemoryBytes: Int64?
-    let minimumSystemMemoryBytes: Int64?
-    let minimumDiskBytes: Int64?
-    let minimumCPUCores: Int?
-    let networkAccess: Bool?
+public struct WorkflowNodeRequirements: Codable, Equatable, Sendable {
+    public let modelIDs: [String]
+    public let acceleratorBackends: [String]
+    public let minimumAcceleratorMemoryBytes: Int64?
+    public let minimumSystemMemoryBytes: Int64?
+    public let minimumDiskBytes: Int64?
+    public let minimumCPUCores: Int?
+    public let networkAccess: Bool?
 
     enum CodingKeys: String, CodingKey {
         case modelIDs = "model_ids"
@@ -391,7 +403,7 @@ struct WorkflowNodeRequirements: Codable, Equatable, Sendable {
         case networkAccess = "network_access"
     }
 
-    init(
+    public init(
         modelIDs: [String],
         acceleratorBackends: [String],
         minimumAcceleratorMemoryBytes: Int64?,
@@ -409,7 +421,7 @@ struct WorkflowNodeRequirements: Codable, Equatable, Sendable {
         self.networkAccess = networkAccess
     }
 
-    init(from decoder: Decoder) throws {
+    public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         modelIDs = try container.decode([String].self, forKey: .modelIDs)
         acceleratorBackends = try container.decode([String].self, forKey: .acceleratorBackends)
@@ -423,7 +435,7 @@ struct WorkflowNodeRequirements: Codable, Equatable, Sendable {
         networkAccess = try container.decodeIfPresent(Bool.self, forKey: .networkAccess)
     }
 
-    static let none = WorkflowNodeRequirements(
+    public static let none = WorkflowNodeRequirements(
         modelIDs: [],
         acceleratorBackends: [],
         minimumAcceleratorMemoryBytes: nil,
@@ -434,12 +446,12 @@ struct WorkflowNodeRequirements: Codable, Equatable, Sendable {
     )
 }
 
-struct WorkflowNodeTraits: Codable, Equatable, Sendable {
-    let deterministic: Bool
-    let cacheable: Bool
-    let sideEffects: String
-    let supportsProgress: Bool
-    let supportsPreviews: Bool
+public struct WorkflowNodeTraits: Codable, Equatable, Sendable {
+    public let deterministic: Bool
+    public let cacheable: Bool
+    public let sideEffects: String
+    public let supportsProgress: Bool
+    public let supportsPreviews: Bool
 
     enum CodingKeys: String, CodingKey {
         case deterministic
@@ -449,7 +461,7 @@ struct WorkflowNodeTraits: Codable, Equatable, Sendable {
         case supportsPreviews = "supports_previews"
     }
 
-    static let core = WorkflowNodeTraits(
+    public static let core = WorkflowNodeTraits(
         deterministic: false,
         cacheable: true,
         sideEffects: "local",
@@ -457,7 +469,7 @@ struct WorkflowNodeTraits: Codable, Equatable, Sendable {
         supportsPreviews: false
     )
 
-    static let material = WorkflowNodeTraits(
+    public static let material = WorkflowNodeTraits(
         deterministic: true,
         cacheable: true,
         sideEffects: "none",
@@ -466,34 +478,34 @@ struct WorkflowNodeTraits: Codable, Equatable, Sendable {
     )
 }
 
-struct WorkflowNodePresentation: Codable, Equatable, Sendable {
-    let style: String
-    let primaryArgument: String?
+public struct WorkflowNodePresentation: Codable, Equatable, Sendable {
+    public let style: String
+    public let primaryArgument: String?
 
     enum CodingKeys: String, CodingKey {
         case style
         case primaryArgument = "primary_argument"
     }
 
-    init(style: String, primaryArgument: String? = nil) {
+    public init(style: String, primaryArgument: String? = nil) {
         self.style = style
         self.primaryArgument = primaryArgument
     }
 }
 
-struct WorkflowNodeCatalogEntry: Codable, Equatable, Sendable {
-    let kind: String
-    let title: String
-    let description: String
-    let category: String
-    let inputs: [WorkflowNodeField]
-    let outputs: [WorkflowNodeOutput]
-    let requirements: WorkflowNodeRequirements
-    let traits: WorkflowNodeTraits
-    let presentation: WorkflowNodePresentation?
-    let provider: WorkflowNodeProviderIdentity?
+public struct WorkflowNodeCatalogEntry: Codable, Equatable, Sendable {
+    public let kind: String
+    public let title: String
+    public let description: String
+    public let category: String
+    public let inputs: [WorkflowNodeField]
+    public let outputs: [WorkflowNodeOutput]
+    public let requirements: WorkflowNodeRequirements
+    public let traits: WorkflowNodeTraits
+    public let presentation: WorkflowNodePresentation?
+    public let provider: WorkflowNodeProviderIdentity?
 
-    init(
+    public init(
         kind: String,
         title: String,
         description: String = "",
@@ -518,8 +530,8 @@ struct WorkflowNodeCatalogEntry: Codable, Equatable, Sendable {
     }
 }
 
-enum WorkflowNodeRegistry {
-    static let entries: [WorkflowNodeCatalogEntry] = [
+public enum WorkflowNodeRegistry {
+    public static let entries: [WorkflowNodeCatalogEntry] = [
         WorkflowNodeCatalogEntry(
             kind: "text.value",
             title: "Text",
@@ -923,11 +935,11 @@ enum WorkflowNodeRegistry {
         ),
     ]
 
-    static func entry(for kind: String) -> WorkflowNodeCatalogEntry? {
+    public static func entry(for kind: String) -> WorkflowNodeCatalogEntry? {
         entries.first { $0.kind == kind }
     }
 
-    static var builtInProvider: WorkflowNodeProviderIdentity {
+    public static var builtInProvider: WorkflowNodeProviderIdentity {
         WorkflowNodeProviderIdentity(
             id: WorkflowNodeProviderIdentity.builtInID,
             version: MereRunCLIVersion.current,
@@ -935,11 +947,7 @@ enum WorkflowNodeRegistry {
         )
     }
 
-    static var catalogEntries: [WorkflowNodeCatalogEntry] {
-        catalogEntries(pluginNodes: WorkflowGraphProviderRegistry.discoveredCatalog().nodes)
-    }
-
-    static func catalogEntries(pluginNodes: [WorkflowNodeCatalogEntry]) -> [WorkflowNodeCatalogEntry] {
+    public static func catalogEntries(pluginNodes: [WorkflowNodeCatalogEntry]) -> [WorkflowNodeCatalogEntry] {
         let builtIn = entries.map { entry in
             WorkflowNodeCatalogEntry(
                 kind: entry.kind,
@@ -959,37 +967,31 @@ enum WorkflowNodeRegistry {
         }
     }
 
-    static func entry(for node: WorkflowNode) -> WorkflowNodeCatalogEntry? {
+    public static func entry(for node: WorkflowNode, pluginNodes: [WorkflowNodeCatalogEntry]) -> WorkflowNodeCatalogEntry? {
         if node.resolvedProviderID == WorkflowNodeProviderIdentity.builtInID {
             return entry(for: node.kind)
         }
-        return WorkflowGraphProviderRegistry.discoveredCatalog().nodes.first {
+        return pluginNodes.first {
             $0.kind == node.kind && $0.provider?.id == node.resolvedProviderID
         }
     }
 
-    static func output(node: WorkflowNode, name: String) -> WorkflowNodeOutput? {
-        entry(for: node)?.outputs.first { $0.name == name }
+    public static func output(node: WorkflowNode, name: String, pluginNodes: [WorkflowNodeCatalogEntry]) -> WorkflowNodeOutput? {
+        entry(for: node, pluginNodes: pluginNodes)?.outputs.first { $0.name == name }
     }
 
-    static func provider(for node: WorkflowNode) -> WorkflowNodeProviderIdentity? {
-        if node.resolvedProviderID == WorkflowNodeProviderIdentity.builtInID {
-            return builtInProvider
-        }
-        return WorkflowGraphProviderRegistry.discoveredCatalog().provider(id: node.resolvedProviderID)?.identity
-    }
 }
 
-struct WorkflowReference: Equatable, Sendable {
-    enum Source: Equatable, Sendable {
+public struct WorkflowReference: Equatable, Sendable {
+    public enum Source: Equatable, Sendable {
         case input(String)
         case nodeOutput(nodeID: String, output: String)
     }
 
-    let rawValue: String
-    let source: Source
+    public let rawValue: String
+    public let source: Source
 
-    init(_ rawValue: String) throws {
+    public init(_ rawValue: String) throws {
         self.rawValue = rawValue
         let parts = rawValue.split(separator: ".", omittingEmptySubsequences: false).map(String.init)
         if parts.count == 2, parts[0] == "inputs" {
@@ -1000,28 +1002,29 @@ struct WorkflowReference: Equatable, Sendable {
             source = .nodeOutput(nodeID: parts[1], output: parts[3])
             return
         }
-        throw ValidationError("Invalid workflow reference '\(rawValue)'.")
+        throw RelayClientError("Invalid workflow reference '\(rawValue)'.")
     }
 }
 
-func workflowSecretEnvironmentKey(_ name: String) -> String {
+public func workflowSecretEnvironmentKey(_ name: String) -> String {
     "MERERUN_SECRET_\(name.uppercased().replacingOccurrences(of: "-", with: "_"))"
 }
 
-struct WorkflowGraphValidation: Equatable {
-    let diagnostics: [PreflightDiagnostic]
-    let order: [String]
-    let dependencies: [String: Set<String>]
+public struct WorkflowGraphValidation: Equatable {
+    public let diagnostics: [PreflightDiagnostic]
+    public let order: [String]
+    public let dependencies: [String: Set<String>]
 
-    var status: StructuredRunStatus {
-        StructuredRunOutput.status(for: diagnostics)
+    public var status: StructuredRunStatus {
+        .forDiagnostics(diagnostics)
     }
 }
 
-enum WorkflowGraphValidator {
-    static func validate(
+public enum WorkflowGraphValidator {
+    public static func validate(
         graph: WorkflowGraphDocument,
-        inputs: WorkflowInputsDocument
+        inputs: WorkflowInputsDocument,
+        pluginNodes: [WorkflowNodeCatalogEntry]
     ) -> WorkflowGraphValidation {
         var diagnostics: [PreflightDiagnostic] = []
         validateHeader(graph, diagnostics: &diagnostics)
@@ -1033,7 +1036,7 @@ enum WorkflowGraphValidator {
         }
         var dependencies: [String: Set<String>] = [:]
         for node in graph.nodes {
-            validateNode(node, graph: graph, nodesByID: nodesByID, diagnostics: &diagnostics)
+            validateNode(node, graph: graph, nodesByID: nodesByID, pluginNodes: pluginNodes, diagnostics: &diagnostics)
             var nodeDependencies = Set(node.dependsOn ?? [])
             for reference in node.arguments.values.flatMap(\.references) {
                 guard let parsed = try? WorkflowReference(reference) else {
@@ -1046,7 +1049,7 @@ enum WorkflowGraphValidator {
             }
             dependencies[node.id] = nodeDependencies
         }
-        validateGraphOutputs(graph, nodesByID: nodesByID, diagnostics: &diagnostics)
+        validateGraphOutputs(graph, nodesByID: nodesByID, pluginNodes: pluginNodes, diagnostics: &diagnostics)
         let order = topologicalOrder(graph.nodes.map(\.id), dependencies: dependencies)
         if order.count != graph.nodes.count {
             diagnostics.append(
@@ -1190,9 +1193,10 @@ enum WorkflowGraphValidator {
         _ node: WorkflowNode,
         graph: WorkflowGraphDocument,
         nodesByID: [String: WorkflowNode],
+        pluginNodes: [WorkflowNodeCatalogEntry],
         diagnostics: inout [PreflightDiagnostic]
     ) {
-        guard let entry = WorkflowNodeRegistry.entry(for: node) else {
+        guard let entry = WorkflowNodeRegistry.entry(for: node, pluginNodes: pluginNodes) else {
             diagnostics.append(.init(
                 id: "workflow_node_kind_unsupported_\(node.id)",
                 severity: .blocker,
@@ -1263,6 +1267,7 @@ enum WorkflowGraphValidator {
                 field: field.name,
                 graph: graph,
                 nodesByID: nodesByID,
+                pluginNodes: pluginNodes,
                 diagnostics: &diagnostics
             )
         }
@@ -1337,6 +1342,7 @@ enum WorkflowGraphValidator {
         field: String,
         graph: WorkflowGraphDocument,
         nodesByID: [String: WorkflowNode],
+        pluginNodes: [WorkflowNodeCatalogEntry],
         diagnostics: inout [PreflightDiagnostic]
     ) {
         if case .reference(let rawReference) = value {
@@ -1360,7 +1366,7 @@ enum WorkflowGraphValidator {
                 }
             case .nodeOutput(let sourceNodeID, let outputName):
                 guard let sourceNode = nodesByID[sourceNodeID],
-                      let output = WorkflowNodeRegistry.output(node: sourceNode, name: outputName) else {
+                      let output = WorkflowNodeRegistry.output(node: sourceNode, name: outputName, pluginNodes: pluginNodes) else {
                     diagnostics.append(invalidReference(rawReference, nodeID: nodeID))
                     return
                 }
@@ -1387,6 +1393,7 @@ enum WorkflowGraphValidator {
                     field: field,
                     graph: graph,
                     nodesByID: nodesByID,
+                    pluginNodes: pluginNodes,
                     diagnostics: &diagnostics
                 )
             }
@@ -1406,6 +1413,7 @@ enum WorkflowGraphValidator {
                         field: field,
                         graph: graph,
                         nodesByID: nodesByID,
+                        pluginNodes: pluginNodes,
                         diagnostics: &diagnostics
                     )
                 }
@@ -1421,6 +1429,7 @@ enum WorkflowGraphValidator {
                         field: field,
                         graph: graph,
                         nodesByID: nodesByID,
+                        pluginNodes: pluginNodes,
                         diagnostics: &diagnostics
                     )
                 }
@@ -1435,6 +1444,7 @@ enum WorkflowGraphValidator {
                     field: field,
                     graph: graph,
                     nodesByID: nodesByID,
+                    pluginNodes: pluginNodes,
                     diagnostics: &diagnostics
                 )
             }
@@ -1457,6 +1467,7 @@ enum WorkflowGraphValidator {
         field: String,
         graph: WorkflowGraphDocument,
         nodesByID: [String: WorkflowNode],
+        pluginNodes: [WorkflowNodeCatalogEntry],
         diagnostics: inout [PreflightDiagnostic]
     ) {
         if case .reference = value {
@@ -1478,6 +1489,7 @@ enum WorkflowGraphValidator {
                 field: field,
                 graph: graph,
                 nodesByID: nodesByID,
+                pluginNodes: pluginNodes,
                 diagnostics: &diagnostics
             )
             return
@@ -1506,6 +1518,7 @@ enum WorkflowGraphValidator {
                     field: "\(field)/\(index)",
                     graph: graph,
                     nodesByID: nodesByID,
+                    pluginNodes: pluginNodes,
                     diagnostics: &diagnostics
                 )
             }
@@ -1520,6 +1533,7 @@ enum WorkflowGraphValidator {
                     field: "\(field)/\(key)",
                     graph: graph,
                     nodesByID: nodesByID,
+                    pluginNodes: pluginNodes,
                     diagnostics: &diagnostics
                 )
             }
@@ -1536,6 +1550,7 @@ enum WorkflowGraphValidator {
     private static func validateGraphOutputs(
         _ graph: WorkflowGraphDocument,
         nodesByID: [String: WorkflowNode],
+        pluginNodes: [WorkflowNodeCatalogEntry],
         diagnostics: inout [PreflightDiagnostic]
     ) {
         for (name, value) in graph.outputs {
@@ -1543,7 +1558,7 @@ enum WorkflowGraphValidator {
                   let reference = try? WorkflowReference(rawReference),
                   case .nodeOutput(let nodeID, let outputName) = reference.source,
                   let node = nodesByID[nodeID],
-                  WorkflowNodeRegistry.output(node: node, name: outputName) != nil else {
+                  WorkflowNodeRegistry.output(node: node, name: outputName, pluginNodes: pluginNodes) != nil else {
                 diagnostics.append(.init(
                     id: "workflow_output_invalid_\(name)",
                     severity: .blocker,
