+++ Tests/MereRunCLITests/RunCommandTests.swift
@@ -1,4 +1,5 @@
 import Foundation
+import MereRunRelayKit
 import MereRunCore
 import XCTest
 @testable import MereRunCLI
